My rmd for questions 1 and 6
This should include the rmd for question 1 and 6. The only thing that the final compiler will need to change is their own code for calling in the data. I tried figuring out how to store it in github and have the code bring it in from there, but I was unsuccessful... so if any of you figure out how to do that, please let me know! :)